fix: location details on new booker (#9710)
* fix: location details on new booker


* chore


---------
